# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for Cemetery API server",
        "title": "Cemetery API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "mailto:support@cemetery.com",
            "email": "support@cemetery.com"
        },
        "version": "0.2"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
        "/api/forgot-password": {
            "post": {
                "description": "Инициирует процесс восстановления пароля пользователя. Отправляет письмо с инструкциями и ссылкой на сброс пароля.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Запрос на восстановление пароля",
                "parameters": [
                    {
                        "description": "Email пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Запрос на восстановление пароля отправлен успешно",
                        "schema": {
                            "$ref": "#/definitions/swagger.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "description": "Обновляет пароль пользователя после сброса. Требуется валидный токен сброса пароля.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Сброс пароля",
                "parameters": [
                    {
                        "description": "Данные для сброса пароля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пароль успешно сброшен",
                        "schema": {
                            "$ref": "#/definitions/swagger.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Токен сброса пароля не найден или устарел",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/time-capsule/create": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Создает капсулу времени для пользователя",
                "tags": [
                    "time-capsule"
                ],
                "summary": "Создание капсулы времени",
                "parameters": [
                    {
                        "description": "Данные для создания капсулы времени",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTimeCapsuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time capsule created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create time capsule",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/time-capsules": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает список капсул времени для пользователя",
                "tags": [
                    "time-capsule"
                ],
                "summary": "Получение списка капсул времени пользователя",
                "responses": {
                    "200": {
                        "description": "Get user's time capsules successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TimeCapsule"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get user's time capsules",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/time-capsule/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает страницу капсулы времени для пользователя",
                "tags": [
                    "time-capsule"
                ],
                "summary": "Получение страницы капсулы времени",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID капсулы времени",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get time capsule page successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get time capsule page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.ContentCategory": {
            "type": "string",
            "enum": [
                "MemoryPage",
                "FamilyTree",
                "Other",
                "TimeCapsule",
                "Letter",
                "VideoMessage"
            ],
            "x-enum-varnames": [
                "MemoryPage",
                "FamilyTree",
                "Other",
                "TimeCapsuleCategory",
                "Letter",
                "VideoMessage"
            ]
        },
...
        "models.CreateTimeCapsuleRequest": {
            "type": "object",
            "required": [
                "category",
                "content",
                "openAt",
                "signature"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "openAt": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "models.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "newPassword",
                "token"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "minLength": 3
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "models.TimeCapsule": {
            "type": "object",
            "properties": {
                "attachedFiles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.ContentCategory"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "openAt": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "swagger.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        }
    }
}